requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.1.0
captcha==0.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
cryptography==46.0.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
extra-streamlit-components==0.1.81
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.31.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.48.0
streamlit-authenticator==0.4.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
